pyproject.toml
[tool.poetry]
name = "django-ninja-test"
version = "0.1.0"
description = ""
authors = ["Marco Peyfuss <mpeyfuss@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
django-ninja = "^1.1.0"
uvicorn = "^0.29.0"
gunicorn = "^22.0.0"
httpx = "^0.27.0"
gevent = "^24.2.1"
django-cors-headers = "^4.3.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.3"
locust = "^2.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
extend-select = ["I"] # enable import sorting